+++ picomint-guardian-daemon/src/ui/dashboard/general.rs
@@ -13,7 +13,7 @@ pub fn render(
     federation_name: &str,
     guardian_names: &BTreeMap<PeerId, String>,
     p2p_connection_status: &BTreeMap<PeerId, P2PConnectionStatus>,
-    session_count: u64,
+    session_count: u32,
     block_count: u32,
     version: ConsensusVersion,
 ) -> Markup {
